+++ Changes.md
@@ -35,6 +35,7 @@ Improvements
 - Backdrop : Improved drawing order for nested backdrops :
   - Larger backdrops are automatically drawn behind smaller ones, so that nested backdrops will always appear on top.
   - Added a `depth` plug to assign a manual drawing depth for the rare cases where the automatic depth is unwanted.
+- ImageStats : Added `areaSource` plug, allowing area to be driven by the input display window or data window.
 
 Fixes
 -----